# api.yaml
openapi: 3.0.0
servers:
  - description: SwaggerHub API Auto Mocking
    url: http://localhost:3000/api
info:
  description: This is a simple API
  version: "1.0.0"
  title: Warehouse management system API
  contact:
    email: you@your-company.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
paths:
...
  /images:
    post:
      tags:
        - images
      summary: Upload ảnh
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Image'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
  /images/{image_id}:
    get:
      tags:
        - images
      summary: Get an image
      parameters:
        - in: path
          name: image_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  img:
                    type: string
                    format: base64
components:
  schemas:
...
    Image:
      type: object
      required:
        - id
        - img_buffer
        - img_type
      properties:
        id:
          type: string
          format: uuid
          example: d290f1ee-6c54-4b01-90e6-d701748f0851
        img_buffer:
          type: object
        img_type:
          type: string
          enum: ['png', 'jpeg', 'jpg', 'webp']